Day 23 in progress

+++ src/main/scala/challenge/Day15.scala
@@ -32,7 +32,7 @@ object Day15 extends Challenge {
 
     @tailrec
     def acc(in: ic.Input, curr: Square, grid: Map[Square, Int]): Map[Square, Int] = {
-      val nextSquare: Square = curr.next(in.in.head)
+      val nextSquare: Square = curr.next(in.in.head.toInt)
       ic.execute(in) match {
         case ic.Output(sig, ptr, rb, state) => // input wanted
           val newGrid = grid.updated(nextSquare, sig.head.toInt)
